==> notebook/1.Extracción con Selenium.ipynb <==
import time
import pandas as pd
import warnings
warnings.filterwarnings('ignore')

from selenium import webdriver
from webdriver_manager.chrome import ChromeDriverManager
from selenium.webdriver.common.by import By
from selenium.webdriver.support.ui import Select

driver = webdriver.Chrome()

# Extraccion de datos Tripadvisor.es:

Comenzamos definiedo la funcion de extraccion.

Definimos las url de las que queremos extrar la informacion ya seleccionada y la guardamos en una lista, con la que luego nos haremos nuestro DF.

def extraer(url):

    global tabla

    driver = webdriver.Chrome()
    driver.get(url)

    time.sleep(2)
    
    # COOKIES

    try:

        aceptar = driver.find_element(By.XPATH, '//*[@id="onetrust-accept-btn-handler"]')
        aceptar.click()

    except:

        print('Ya')

    time.sleep(5)

    tarjetas = []

    for e in range(len(driver.find_elements(By.TAG_NAME, 'header'))):
        box = driver.find_elements(By.TAG_NAME, 'header')[e].text.split('\n')
        tarjetas.append(box)
        
    tabla += tarjetas


url1='https://www.tripadvisor.es/FindRestaurants?geo=187514&cuisines=10749%2C10639%2C10659%2C4617%2C5473%2C10653%2C5110%2C5379%2C10698%2C10631%2C10715%2C10772%2C5086%2C20076%2C10687%2C10346%2C10660%2C20062%2C10622%2C10757%2C10661%2C20069%2C10348%2C10347%2C10626%2C10791%2C10633%2C10692%2C20066%2C10680%2C10744%2C10632%2C10752%2C20068%2C20067%2C11744%2C10618%2C11739%2C10769%2C10742%2C10765%2C10753%2C10732%2C10635%2C10767%2C10662%2C10693%2C10695%2C10664%2C10636%2C10617%2C10628%2C21363%2C10675%2C20074%2C10696%2C10748%2C10756%2C10690%2C10681%2C10717%2C10747%2C10784%2C10775%2C10766%2C10755&establishmentTypes=10591&broadened=false'

url2='https://www.tripadvisor.es/FindRestaurants?geo=187514&offset=30&cuisines=10749%2C10639%2C10659%2C4617%2C5473%2C10653%2C5110%2C5379%2C10698%2C10631%2C10715%2C10772%2C5086%2C20076%2C10687%2C10346%2C10660%2C20062%2C10622%2C10757%2C10661%2C20069%2C10348%2C10347%2C10626%2C10791%2C10633%2C10692%2C20066%2C10680%2C10744%2C10632%2C10752%2C20068%2C20067%2C11744%2C10618%2C11739%2C10769%2C10742%2C10765%2C10753%2C10732%2C10635%2C10767%2C10662%2C10693%2C10695%2C10664%2C10636%2C10617%2C10628%2C21363%2C10675%2C20074%2C10696%2C10748%2C10756%2C10690%2C10681%2C10717%2C10747%2C10784%2C10775%2C10766%2C10755&establishmentTypes=10591&broadened=false'

url3='https://www.tripadvisor.es/FindRestaurants?geo=187514&offset=60&cuisines=10749%2C10639%2C10659%2C4617%2C5473%2C10653%2C5110%2C5379%2C10698%2C10631%2C10715%2C10772%2C5086%2C20076%2C10687%2C10346%2C10660%2C20062%2C10622%2C10757%2C10661%2C20069%2C10348%2C10347%2C10626%2C10791%2C10633%2C10692%2C20066%2C10680%2C10744%2C10632%2C10752%2C20068%2C20067%2C11744%2C10618%2C11739%2C10769%2C10742%2C10765%2C10753%2C10732%2C10635%2C10767%2C10662%2C10693%2C10695%2C10664%2C10636%2C10617%2C10628%2C21363%2C10675%2C20074%2C10696%2C10748%2C10756%2C10690%2C10681%2C10717%2C10747%2C10784%2C10775%2C10766%2C10755&establishmentTypes=10591&broadened=false'

url4='https://www.tripadvisor.es/FindRestaurants?geo=187514&offset=90&cuisines=10749%2C10639%2C10659%2C4617%2C5473%2C10653%2C5110%2C5379%2C10698%2C10631%2C10715%2C10772%2C5086%2C20076%2C10687%2C10346%2C10660%2C20062%2C10622%2C10757%2C10661%2C20069%2C10348%2C10347%2C10626%2C10791%2C10633%2C10692%2C20066%2C10680%2C10744%2C10632%2C10752%2C20068%2C20067%2C11744%2C10618%2C11739%2C10769%2C10742%2C10765%2C10753%2C10732%2C10635%2C10767%2C10662%2C10693%2C10695%2C10664%2C10636%2C10617%2C10628%2C21363%2C10675%2C20074%2C10696%2C10748%2C10756%2C10690%2C10681%2C10717%2C10747%2C10784%2C10775%2C10766%2C10755&establishmentTypes=10591&broadened=false'

url5='https://www.tripadvisor.es/FindRestaurants?geo=187514&offset=90&cuisines=10749%2C10639%2C10659%2C4617%2C5473%2C10653%2C5110%2C5379%2C10698%2C10631%2C10715%2C10772%2C5086%2C20076%2C10687%2C10346%2C10660%2C20062%2C10622%2C10757%2C10661%2C20069%2C10348%2C10347%2C10626%2C10791%2C10633%2C10692%2C20066%2C10680%2C10744%2C10632%2C10752%2C20068%2C20067%2C11744%2C10618%2C11739%2C10769%2C10742%2C10765%2C10753%2C10732%2C10635%2C10767%2C10662%2C10693%2C10695%2C10664%2C10636%2C10617%2C10628%2C21363%2C10675%2C20074%2C10696%2C10748%2C10756%2C10690%2C10681%2C10717%2C10747%2C10784%2C10775%2C10766%2C10755&establishmentTypes=10591&broadened=false'

url6 = 'https://www.tripadvisor.es/FindRestaurants?geo=187514&offset=120&cuisines=10749%2C10639%2C10659%2C4617%2C5473%2C10653%2C5110%2C5379%2C10698%2C10631%2C10715%2C10772%2C5086%2C20076%2C10687%2C10346%2C10660%2C20062%2C10622%2C10757%2C10661%2C20069%2C10348%2C10347%2C10626%2C10791%2C10633%2C10692%2C20066%2C10680%2C10744%2C10632%2C10752%2C20068%2C20067%2C11744%2C10618%2C11739%2C10769%2C10742%2C10765%2C10753%2C10732%2C10635%2C10767%2C10662%2C10693%2C10695%2C10664%2C10636%2C10617%2C10628%2C21363%2C10675%2C20074%2C10696%2C10748%2C10756%2C10690%2C10681%2C10717%2C10747%2C10784%2C10775%2C10766%2C10755&establishmentTypes=10591&broadened=false'

url7 = 'https://www.tripadvisor.es/FindRestaurants?geo=187514&offset=150&cuisines=10749%2C10639%2C10659%2C4617%2C5473%2C10653%2C5110%2C5379%2C10698%2C10631%2C10715%2C10772%2C5086%2C20076%2C10687%2C10346%2C10660%2C20062%2C10622%2C10757%2C10661%2C20069%2C10348%2C10347%2C10626%2C10791%2C10633%2C10692%2C20066%2C10680%2C10744%2C10632%2C10752%2C20068%2C20067%2C11744%2C10618%2C11739%2C10769%2C10742%2C10765%2C10753%2C10732%2C10635%2C10767%2C10662%2C10693%2C10695%2C10664%2C10636%2C10617%2C10628%2C21363%2C10675%2C20074%2C10696%2C10748%2C10756%2C10690%2C10681%2C10717%2C10747%2C10784%2C10775%2C10766%2C10755&establishmentTypes=10591&broadened=false'

url8 = 'https://www.tripadvisor.es/FindRestaurants?geo=187514&offset=180&cuisines=10749%2C10639%2C10659%2C4617%2C5473%2C10653%2C5110%2C5379%2C10698%2C10631%2C10715%2C10772%2C5086%2C20076%2C10687%2C10346%2C10660%2C20062%2C10622%2C10757%2C10661%2C20069%2C10348%2C10347%2C10626%2C10791%2C10633%2C10692%2C20066%2C10680%2C10744%2C10632%2C10752%2C20068%2C20067%2C11744%2C10618%2C11739%2C10769%2C10742%2C10765%2C10753%2C10732%2C10635%2C10767%2C10662%2C10693%2C10695%2C10664%2C10636%2C10617%2C10628%2C21363%2C10675%2C20074%2C10696%2C10748%2C10756%2C10690%2C10681%2C10717%2C10747%2C10784%2C10775%2C10766%2C10755&establishmentTypes=10591&broadened=false'

url9 = 'https://www.tripadvisor.es/FindRestaurants?geo=187514&offset=180&cuisines=10749%2C10639%2C10659%2C4617%2C5473%2C10653%2C5110%2C5379%2C10698%2C10631%2C10715%2C10772%2C5086%2C20076%2C10687%2C10346%2C10660%2C20062%2C10622%2C10757%2C10661%2C20069%2C10348%2C10347%2C10626%2C10791%2C10633%2C10692%2C20066%2C10680%2C10744%2C10632%2C10752%2C20068%2C20067%2C11744%2C10618%2C11739%2C10769%2C10742%2C10765%2C10753%2C10732%2C10635%2C10767%2C10662%2C10693%2C10695%2C10664%2C10636%2C10617%2C10628%2C21363%2C10675%2C20074%2C10696%2C10748%2C10756%2C10690%2C10681%2C10717%2C10747%2C10784%2C10775%2C10766%2C10755&establishmentTypes=10591&broadened=false'

url10 = 'https://www.tripadvisor.es/FindRestaurants?geo=187514&offset=210&cuisines=10749%2C10639%2C10659%2C4617%2C5473%2C10653%2C5110%2C5379%2C10698%2C10631%2C10715%2C10772%2C5086%2C20076%2C10687%2C10346%2C10660%2C20062%2C10622%2C10757%2C10661%2C20069%2C10348%2C10347%2C10626%2C10791%2C10633%2C10692%2C20066%2C10680%2C10744%2C10632%2C10752%2C20068%2C20067%2C11744%2C10618%2C11739%2C10769%2C10742%2C10765%2C10753%2C10732%2C10635%2C10767%2C10662%2C10693%2C10695%2C10664%2C10636%2C10617%2C10628%2C21363%2C10675%2C20074%2C10696%2C10748%2C10756%2C10690%2C10681%2C10717%2C10747%2C10784%2C10775%2C10766%2C10755&establishmentTypes=10591&broadened=false'

url11= 'https://www.tripadvisor.es/FindRestaurants?geo=187514&offset=240&cuisines=10749%2C10639%2C10659%2C4617%2C5473%2C10653%2C5110%2C5379%2C10698%2C10631%2C10715%2C10772%2C5086%2C20076%2C10687%2C10346%2C10660%2C20062%2C10622%2C10757%2C10661%2C20069%2C10348%2C10347%2C10626%2C10791%2C10633%2C10692%2C20066%2C10680%2C10744%2C10632%2C10752%2C20068%2C20067%2C11744%2C10618%2C11739%2C10769%2C10742%2C10765%2C10753%2C10732%2C10635%2C10767%2C10662%2C10693%2C10695%2C10664%2C10636%2C10617%2C10628%2C21363%2C10675%2C20074%2C10696%2C10748%2C10756%2C10690%2C10681%2C10717%2C10747%2C10784%2C10775%2C10766%2C10755&establishmentTypes=10591&broadened=false'


tabla=[]

extraer(url1)
extraer(url2)
extraer(url3)
extraer(url4)
extraer(url5)
extraer(url6)
extraer(url7)
extraer(url8)
extraer(url9)
extraer(url10)


data=pd.DataFrame(tabla)

data.to_json('../data/Tripadvisor2.json', index=False)


data.tail()
