==> notebook/2.APIgoogle.ipynb <==
import requests
import json
import pandas as pd

import sys
sys.path.append("..")
from src.funciones import *

# Descarga tablas por tipo de restaurante y sus opiniones:

Tendremos los datos de los restaurantes y el numero de opiniones. Solo me descargo los tipos de comida que me aparecene en el top según Tripadvisor, pero luego tendré que comprobar si coinciden.

with open('API.txt') as file:
    
    key = file.read() #cargo mi contraseña y la guardo en una variable

japo = pd.DataFrame(consulta_API('japones'))
japo = japo.drop(columns=['icon','icon_background_color','icon_mask_base_uri','opening_hours','place_id','plus_code','reference','scope','vicinity'])
japo.to_csv('../data/japo.csv',index=False)

chino = pd.DataFrame(consulta_API('Chino'))
chino = chino.drop(columns=['icon','icon_background_color','icon_mask_base_uri','opening_hours','place_id','plus_code','reference','scope','vicinity'])
chino.to_csv('../data/chino.csv',index=False)

coreano = pd.DataFrame(consulta_API('Coreano'))
coreano = coreano.drop(columns=['icon','icon_background_color','icon_mask_base_uri','opening_hours','place_id','plus_code','reference','scope','vicinity'])
coreano.to_csv('../data/coreano.csv',index=False)

italiano = pd.DataFrame(consulta_API('Italiano'))
italiano = italiano.drop(columns=['icon','icon_background_color','icon_mask_base_uri','opening_hours','place_id','plus_code','reference','scope','vicinity'])
italiano.to_csv('../data/ita.csv',index=False)

cari = pd.DataFrame(consulta_API('Caribeña'))
cari = cari.drop(columns=['icon','icon_background_color','icon_mask_base_uri','opening_hours','place_id','plus_code','reference','scope','vicinity'])
cari.to_csv('../data/caribe.csv',index=False)

indi = pd.DataFrame(consulta_API('India'))
indi = indi.drop(columns=['icon','icon_background_color','icon_mask_base_uri','opening_hours','place_id','plus_code','reference','scope','vicinity'])
indi.to_csv('../data/india.csv',index=False)

mex = pd.DataFrame(consulta_API('Mexicana'))
mex = mex.drop(columns=['icon','icon_background_color','icon_mask_base_uri','opening_hours','place_id','plus_code','reference','scope','vicinity'])
mex.to_csv('../data/mex.csv',index=False)

peru = pd.DataFrame(consulta_API('Peruana'))
peru = peru.drop(columns=['icon','icon_background_color','icon_mask_base_uri','opening_hours','place_id','plus_code','reference','scope','vicinity'])
peru.to_csv('../data/peru.csv',index=False)
