==> notebook/3.Limpieza_json.ipynb <==
import time
import pandas as pd
import warnings
warnings.filterwarnings('ignore')

import asyncio

from selenium import webdriver
from webdriver_manager.chrome import ChromeDriverManager
from selenium.webdriver.common.by import By
from selenium.webdriver.support.ui import Select


# Descargamos el json con los datos que queremos:

import requests

url = 'https://datos.madrid.es/egob/catalogo/209548-491-censo-locales-historico.json'  

response = requests.get(url)

if response.status_code == 200:
    
    data = response.json()
    df = pd.DataFrame(data)

else:
    print("No se pudo descargar el archivo JSON.")


data=df.copy()

data.info()

data.head()

#Comnezamos eliminando las columnas inecesarias

data = data.drop(columns=['hora_cierre2','hora_apertura2','fx_carga','cal_acceso','num_acceso','nom_acceso','id_seccion_censal_local','desc_seccion_censal_local','coordenada_x_agrupacion','coordenada_y_agrupacion','hora_cierre1','hora_apertura1'])

#data = data.drop(columns=['id_tipo_acceso_local','desc_tipo_acceso_local','id_agrupacion'])
#data = data.drop(columns=['id_vial_edificio','clase_vial_edificio'])
#data = data.drop(columns=['id_situacion_local','desc_situacion_local','id_clase_ndp_acceso','nombre_agrupacion','id_tipo_agrup','desc_tipo_agrup','id_planta_agrupado','id_local_agrupado','fx_datos_ini','fx_datos_fin'])

data = data.drop(columns=['id_situacion_local','desc_situacion_local','id_clase_ndp_acceso','nombre_agrupacion','id_tipo_agrup','desc_tipo_agrup','id_planta_agrupado','id_local_agrupado','fx_datos_ini','fx_datos_fin'])

data.columns

data.reset_index()

data=data.drop(columns=[])

import numpy as np
nueva_columna = np.arange(1, len(df) + 1)
data.id_local=nueva_columna

data.columns

data=data.drop(columns='desc_vial_edificio')

mascara = data['rotulo'] != 'SIN ACTIVIDAD'

# Utiliza la máscara para seleccionar las filas que no coinciden con "SIN ACTIVIDAD"
data = data[mascara]

data.head()

localizacion = ['id_local','id_distrito_local','desc_distrito_local','id_barrio_local', 'desc_barrio_local', 'cod_barrio_local','coordenada_x_local', 'coordenada_y_local']
loc = pd.DataFrame(data, columns=localizacion)
loc

coor=['id_local','coordenada_x_local','coordenada_y_local']
coordenadas=pd.DataFrame(loc, columns=coor)
coordenadas

mascara = (coordenadas['coordenada_x_local'] == 0) & (coordenadas['coordenada_y_local'] == 0)

# Utiliza la máscara para eliminar las filas correspondientes, las que son 0
coordenadas = coordenadas[~mascara]

coordenadas = coordenadas.astype(int)

coordenadas.describe().T

data.columns

rest=['id_local','rotulo']
restaurantes=pd.DataFrame(data, columns=rest)
restaurantes

import sys
sys.path.append("..")
from src.funciones import *

optim(restaurantes)

restaurantes.to_json('../data/restautantes.json')
