==> notebook/5.Limpieza_Tripadvisor.ipynb <==
import pandas as pd

import sys
sys.path.append("..")
from src.funciones import * 

# Limpieza archivo CSV (Fuente: Tripadvisor):

Una vez tenemos el archivos csv,procedemos a limpiarlo.

Tras ver el dataframe procedemos con la siguiente limpieza:

* Ponemos los títulos bien
* Quitamos las filas que han salido mal durante la extracción de datos
* Quitamos los números que preceden a los nomber
* Reseteamos los índices
* Quitamos de la columna opiniones Abierto/Cerrado ahora
* Quitamos de Tipo '€' y los añadimos a precio la equivalencia
* Quitamos Carta de Tipo

ruta_al_csv = '..\data\Tripadvisor2.csv'
df = pd.read_csv(ruta_al_csv)
data=df.copy()

data.info()

data.describe()

data.columns=['Nombre','Opiniones','Tipo','Precio','Cosas'] # renombramos columnas

eliminar=['Patrocinado', 'Descubrir','MICHELIN','Opinión'] # quitamos las filas que se repiten y no nos interesan porque se han colado
data = data[~data['Nombre'].isin(eliminar)]

data['Nombre'] = data['Nombre'].str.replace(r'\d+\.\s', '', regex=True) #eliminamos numeros de nombre
data = data.reset_index(drop=True)

# Limpiamos loa columna Opiniones y Tipo

data['Opiniones'] = data['Opiniones'].str.replace('Abierto ahora', '')
data['Opiniones'] = data['Opiniones'].str.replace('Cerrado ahora', '')

data['Tipo'] = data['Tipo'].str.replace(r'Carta$', '', regex=True)
data['Precio'] = data['Tipo'].str.extract(r'(€+)')


def precio(tipo):
    if '€€€€' in tipo:
        return 'Alto'
    elif '€€ - €€€' in tipo:
        return 'Medio'
    elif '€' in tipo:
        return 'Bajo'
    else:
        return ''
    

data['Precio'] = data['Tipo'].apply(precio)
data['Tipo'] = data['Tipo'].str.replace('€', '', regex=False)
data['Tipo'] = data['Tipo'].str.replace('-', '', regex=False)
data.head()


limpieza_tipo(data)


import regex as re

cerrado = [re.sub(r'Cerrado hoy', '', cadena) for cadena in data.Opiniones]
minutos = [re.sub(r'Abre en 6 minutos', '', opinion) for opinion in data.Opiniones]

data['Opiniones'] = data['Opiniones'].apply(lambda x: re.sub(r'Abre en 6 minutos', '', x))
data['Opiniones'] = data['Opiniones'].apply(lambda x: re.sub(r'Cerrado hoy', '', x))

def eliminar_palabras(texto):
    palabras_a_eliminar = ["Latina", "Francesa", "Saludable", "Internacional",'rápida','saludable']
    patron = r'\b(?:' + '|'.join(palabras_a_eliminar) + r')\b'
    return re.sub(patron, '', texto)

# Aplica la función a la columna "Tipo"

data['Tipo'] = data['Tipo'].apply(eliminar_palabras)


data.head()

valores_unicos = data['Tipo'].value_counts()
valores_unicos = valores_unicos[valores_unicos == 1]

# Convierte los valores únicos en una lista
valores_unicos_lista = valores_unicos.index.tolist()
valores_unicos_lista

eliminar=['LatinaComida   ',
 'LatinaInternacional  ',
 'Comida rápidaInternacional  ',
 'FusiónSaludable  ',]

data.index[data['Tipo'].isin(eliminar)]

data = data.drop(data.index[data['Tipo'].isin(eliminar)])

data.Nombre = data.Nombre.str.upper()

data.info(memory_usage='deep')

data = optim(data)

data.info(memory_usage='deep')

data.to_json('../data/Tripadvisor.json')
