==> notebook/6.Limpieza_API_Google.ipynb <==
import pandas as pd


# Limpieza tablas extraidas desde la API PLACE de Google: 

#### Todas las tablas tienen la misma estructura por lo que la misma limpieza se la aplico a todas.

caribe = pd.read_csv('../data/caribe.csv', encoding='latin1')
coreano = pd.read_csv('../data/coreano.csv', encoding='latin1')
india = pd.read_csv('../data/india.csv', encoding='latin1')
ita = pd.read_csv('../data/ita.csv', encoding='latin1')
japo = pd.read_csv('../data/japo.csv', encoding='latin1')
mex = pd.read_csv('../data/mex.csv', encoding='latin1')
peru = pd.read_csv('../data/peru.csv', encoding='latin1')
chino = pd.read_csv('../data/chino.csv', encoding='latin1')

coreano.info() # Aqui veo que solo price-level tiene nulos

coreano.columns

coreano=coreano.drop(columns=['business_status','geometry','tipo'])

coreano.columns = ['Nombre','Foto','Precio','Puntuacion','Servicios', 'Nº opiniones']

coreano.Nombre = coreano.Nombre.str.upper()

coreano.head() 

#### Creo una función para limpiar 🫧🧹 todo de una.

import sys
sys.path.append("..")
from src.funciones import *  #funciones

india = clean(india) 
ita = clean(ita) 
japo = clean(japo)
mex = clean(mex) 
peru = clean(peru)
chino = clean(chino)

peru

#### La talba de Caribe tiene una columna menos por lo que la funcion no se puede aplicar

caribe=caribe.drop(columns=['business_status','geometry','tipo'])
caribe.columns = ['Nombre','Foto','Puntuacion','Servicios', 'Nº opiniones']
caribe.Nombre = caribe.Nombre.str.upper()

#### Nuesto principal objetivo 🎯🎯 es tener todos los nombres de los restaurantes de las tablas para poder relacionarlos con la tabla de locales

locales = pd.read_csv('../data/restaurantes.csv', encoding='latin1')

locales.columns = ['o', 'id_local','Nombre']

df_list = [caribe, japo, chino, mex, peru, india, ita]


results = []

for df in df_list:
    result = locales.merge(df, on='Nombre', how='inner')
    results.append(result)

# Concatenar todos los resultados en un solo DataFrame
result = pd.concat(results, ignore_index=True)

result = result.drop_duplicates(subset=['id_local'], keep=False)


result= result.drop(columns=['o','Foto'])

result.head()

result.info(memory_usage='deep')

optim(result)
result.info(memory_usage='deep')

result.to_csv('../data/Opiniones.csv',index=False)

india = optim(india) 
ita = optim(ita) 
japo = optim(japo)
mex = optim(mex) 
peru = optim(peru)
chino = optim(chino)
caribe = optim(caribe)

india.to_json('../data/Indio.json',index=False)
ita.to_json('../data/Italiano.json',index=False)
japo.to_json('../data/Japonés.json',index=False)
mex.to_json('../data/Mexicano.json',index=False)
peru.to_json('../data/Peruano.json',index=False)
chino.to_json('../data/Chino.json',index=False)
caribe.to_json('../data/Caribe.json',index=False)
