==> notebook/7.DataBase.ipynb <==
from pymongo import MongoClient
import os
import json

# Subimos nuestros data a Mongo.

cursor=MongoClient('mongodb://localhost:27017')

cursor.list_database_names()

db = MongoClient.restaurantes  
colec = db.r_etnicos

db.list_collection_names()

json=colec.insert_many(json)
